Fix install_dotfiles.sh to properly handle directories with spaces in the name

+++ install_dotfiles.sh
@@ -1,8 +1,13 @@
-#!/bin/bash
+#!/bin/bash -el
 ############################
 # install_dotfiles.sh
 # This script creates symlinks from the home directory to any desired dotfiles in ~/dotfiles
 ############################
+
+log_debug() {
+  [[ -z "${RAD_DEBUG}" ]] && return 0
+  >&2 echo "$@"
+}
 
 # Build list of files to symlink
 # Returns absolute paths
@@ -10,32 +15,36 @@ build_files_list() {
   local dotfiles_global_dir="$1"
   local dotfiles_dir="$2"
 
-  local global_dotfiles="${dotfiles_global_dir}/*"
-  local profile_dotfiles="${dotfiles_dir}/*"
+  local dotfiles_global="${dotfiles_global_dir}"
+  local dotfiles_profile="${dotfiles_dir}"
   local filename
+
+  log_debug "global dotfiles dir: ${dotfiles_global_dir}"
 
   [[ -z $dotfiles_global_dir ]] || [[ -z $dotfiles_dir ]] && { echo "ERROR: Check args to build_files_list"; exit 1; }
 
+#  IFS="\n"
+
   # get global dotfiles
-  for filepath in ${global_dotfiles}; do
+  for filepath in "${dotfiles_global}"/*; do
     # skip markdown files
     [[ "${filepath##*.}" == "md" ]] && continue
 
-    filename="$(basename ${filepath})"
+    filename="$(basename "${filepath}")"
     if [[ -f "${dotfiles_dir}/${filename}" ]]; then
       >&2 echo "skipping global filepath in favor of profile-specific filepath for ${filename}"
       continue
     fi
 
-    echo "${filepath}"
+    GLOBAL_filepaths+=("${filepath}")
   done
 
   # get profile-specific dotfiles
-  for filepath in ${profile_dotfiles}; do
+  for filepath in "${dotfiles_profile}"/*; do
     # skip markdown files
     [[ "${filepath##*.}" == "md" ]] && continue
 
-    echo "${filepath}"
+    GLOBAL_filepaths+=("${filepath}")
   done
 }
 
@@ -43,13 +52,19 @@ create_symlink() {
   local source_file="$1"
   local target_file="$2"
 
+  # Ensure the source file actually exists
+  if [[ ! -f "${source_file}" ]]; then
+    >&2 echo "ERROR: Source file '${source_file}' does not exist"
+    exit 0
+  fi
+
   # Check if the symlink already exists and points to the correct path
   if [[ ! -L "$target_file" || "$(readlink -f "$target_file")" != "$(readlink -f "$source_file")" ]]; then
     # Create or update the symlink
     ln -sf "$source_file" "$target_file"
   else
     # do nothing
-    :
+    log_debug "DEBUG: Ignoring target '${target_file}' (source: ${source_file})"
   fi
 }
 
@@ -74,15 +89,16 @@ if [[ ! -d $dotfiles_dir ]]; then
 fi
 
 # create dotfiles_old in homedir
-echo "Creating $dotfile_backup_dir for backup of any existing dotfiles in \$HOME"
+echo "Creating '$dotfile_backup_dir' for backup of any existing dotfiles in \$HOME"
 mkdir -p $dotfile_backup_dir
 
 # build list of files
-filepaths="$(build_files_list ${dotfiles_global_dir} ${dotfiles_dir})"
+GLOBAL_filepaths=()
+build_files_list "${dotfiles_global_dir}" "${dotfiles_dir}"
 
 # move any existing dotfiles in homedir to dotfiles_old directory
 echo "Moving any existing non-symlink dotfiles from \$HOME to $dotfile_backup_dir"
-for filepath in ${filepaths}; do
+for filepath in "${GLOBAL_filepaths[@]}"; do
   filename="$(basename "${filepath}")"
   if [[ -f "${HOME}/.${filename}" && ! -L "${HOME}/.${filename}" ]]; then
     echo "Backing up existing dotfile: ~/.${filename} to ${dotfile_backup_dir}"
@@ -91,7 +107,7 @@ for filepath in ${filepaths}; do
 done
 
 # create symlinks
-for filepath in ${filepaths}; do
+for filepath in "${GLOBAL_filepaths[@]}"; do
   filename="$(basename "${filepath}")"
   echo "Creating symlink to ${filename} in \$HOME"
   create_symlink "${filepath}" "${HOME}/.${filename}"
